Php Term Project
This was a php term project for my database class.  The entire program
is a command line interface that calls down to a mysql server and make
orders.  The concept was to make a program where students could
fundraise for themselves by having the customer provide the students
name when they use the service.